feat(reservation): car.vue file added
